test: ensure all products are rendered
